* Readonly support for Inputbox * Geometry boxes now 0 initialize contents * If inputbox looses focus when dirty, a change event is raised

+++ Source/Gorgon/Graphics/BitmapFont.h
@@ -229,7 +229,7 @@ namespace Gorgon { namespace Graphics {
 
 		virtual int GetDigitWidth() const override { return digw; }
 
-		virtual float GetLineThickness() const override { return linethickness; }
+		virtual float GetLineThickness() const override { return (float)linethickness; }
 
 		virtual int GetUnderlineOffset() const override { return underlinepos; }
